requirements.txt
tensorflow==2.15.0
fastapi
gradio
h5py
uvicorn
pillow
opencv-python
numpy
gdown
python-multipart